perf: a cleanup is a Bx, so keep it in a Bx's worth of room
Every frame and every coroutine carried 32 size_t cleanups -- 256 bytes
of room for what is stored there, which is the Bx of the PUSHCLEANUP
that entered the block. A Bx is 16 bits (code.h's lhat_bx), and a jump
could not reach past that anyway, so 248 of those bytes were never
going to hold anything. Most coroutines carry no cleanups at all.

Narrowed to uint16_t, with the count in a uint8_t beside it, and static
asserts tying all three to the widths they have to hold, so the room
cannot drift away from what is put in it.

  LhatCoroutine  440 -> 240 bytes
  Frame          648 -> 448
  Machine     204,272 -> 164,272   (200 frames live in it)

+++ include/lhat/object.h
@@ -218,9 +218,13 @@ typedef struct LhatCoroutine {
     size_t pc;
     uint8_t sent_into;     // the register a resume's value arrives in
 
-    // 02 の 10.7: what a discarded coroutine still has to run.
-    size_t cleanups[LHAT_COROUTINE_CLEANUPS];
-    size_t cleanup_count;
+    // 02 の 10.7: what a discarded coroutine still has to run. Each is an
+    // instruction index, which is a Bx and so 16 bits wide wherever it is
+    // written down (code.h's lhat_bx) -- a jump could not reach past that
+    // anyway. Most coroutines carry none of these, so the room the array
+    // costs whether or not it is used is worth keeping small.
+    uint16_t cleanups[LHAT_COROUTINE_CLEANUPS];
+    uint8_t cleanup_count;
 
     // 02 の 10.7: the collector is the last place a discarded coroutine's
     // cleanups can be reached from, so one found unreachable with some still
